Fixed NT Install, finding files, and post-processor.
Rewrote the logic to determine if the NT database was installed. Changed
all find commands to execute locate if it's installed, but if the file
is not found by locate or if locate is not installed, find is run.

Reserialization of the configuration files was failing, so I added a
task which will automatically be executed at the end of each task list
instead. Not fully tested.